+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/MediaSource.java
@@ -30,6 +30,7 @@ import androidx.media3.exoplayer.drm.DrmSessionManagerProvider;
 import androidx.media3.exoplayer.upstream.Allocator;
 import androidx.media3.exoplayer.upstream.CmcdConfiguration;
 import androidx.media3.exoplayer.upstream.LoadErrorHandlingPolicy;
+import androidx.media3.extractor.text.SubtitleParser;
 import java.io.IOException;
 
 /**
@@ -113,6 +114,19 @@ public interface MediaSource {
     @UnstableApi
     default Factory experimentalParseSubtitlesDuringExtraction(
         boolean parseSubtitlesDuringExtraction) {
+      return this;
+    }
+
+    /**
+     * Sets the {@link SubtitleParser.Factory} to be used for parsing subtitles during extraction if
+     * {@link #experimentalParseSubtitlesDuringExtraction} is enabled.
+     *
+     * @param subtitleParserFactory The {@link SubtitleParser.Factory} for parsing subtitles during
+     *     extraction.
+     * @return This factory, for convenience.
+     */
+    @UnstableApi
+    default Factory setSubtitleParserFactory(SubtitleParser.Factory subtitleParserFactory) {
       return this;
     }