+++ libraries/extractor/src/main/java/androidx/media3/extractor/ExtractorsFactory.java
@@ -18,6 +18,7 @@ package androidx.media3.extractor;
 import android.net.Uri;
 import androidx.media3.common.MimeTypes;
 import androidx.media3.common.util.UnstableApi;
+import androidx.media3.extractor.text.SubtitleParser;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.util.List;
 import java.util.Map;
@@ -47,6 +48,19 @@ public interface ExtractorsFactory {
   @CanIgnoreReturnValue
   default ExtractorsFactory experimentalSetTextTrackTranscodingEnabled(
       boolean textTrackTranscodingEnabled) {
+    return this;
+  }
+
+  /**
+   * Sets a {@link SubtitleParser.Factory} to use when transcoding text tracks.
+   *
+   * <p>This is only works if {@link #experimentalSetTextTrackTranscodingEnabled(boolean)} is
+   * enabled.
+   *
+   * @param subtitleParserFactory The factory for {@link SubtitleParser} instances.
+   * @return The factory, for convenience.
+   */
+  default ExtractorsFactory setSubtitleParserFactory(SubtitleParser.Factory subtitleParserFactory) {
     return this;
   }